don't use builtins.fetchGit for cargo git dependencies
Unlike nixpkgs fetchers (such as pkgs.fetchgit or pkgs.fetchFromGitLab),
which are fixed-output derivations and therefore do the actual fetching
only when they're being built; builtins.fetchGit fetches when the
expression is being evaluated (regardless of whether the package for
which it provides the source is even being built).
Also, because it doesn't evaluate to a derivation that can be built, its
result cannot be pushed to a binary cache.

This means that the git repos using builtins.fetchGit need to be fetched
each time a package using them is being evaluated (although there's some
level of local caching), even when the package itself has already been
built.

In contrast, when using nixpkgs fetchers for src, the src attribute is
basically just another derivation the package depends on, and will be
built (or substituted!) when building the package.

It's for these reasons that builtin fetchers such as builtins.fetchGit
aren't being used in nixpkgs.